consolidate log functions
